pyproject.toml
[tool.poetry]
name = "envmanager"
version = "0.1.0"
description = ""
authors = ["mannyuncharted <omoebun52@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.20.0"
gitpython = "^3.1.43"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
